Consolidate schemas
