+++ src/controllers/api/2014/ruleController.ts
@@ -1,19 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
-import { z } from 'zod';
+import { NameDescQuerySchema, ShowParamsSchema } from '@/schemas/schemas';
 import { ResourceList, escapeRegExp, redisClient } from '@/util';
 
 import Rule from '@/models/2014/rule';
-
-// --- Zod Schemas ---
-const IndexQuerySchema = z.object({
-  name: z.string().optional(),
-  desc: z.string().optional(),
-});
-
-const ShowParamsSchema = z.object({
-  index: z.string().min(1),
-});
-// --- End Zod Schemas ---
 
 interface IndexQuery {
   name?: { $regex: RegExp };
@@ -23,7 +12,7 @@ interface IndexQuery {
 export const index = async (req: Request, res: Response, next: NextFunction) => {
   try {
     // Validate query parameters
-    const validatedQuery = IndexQuerySchema.safeParse(req.query);
+    const validatedQuery = NameDescQuerySchema.safeParse(req.query);
     if (!validatedQuery.success) {
       return res
         .status(400)
