+++ src/controllers/api/2014/subraceController.ts
@@ -1,16 +1,10 @@
 import { Request, Response, NextFunction } from 'express';
-import { z } from 'zod';
+import { ShowParamsSchema } from '@/schemas/schemas';
 import Proficiency from '@/models/2014/proficiency';
 import { ResourceList } from '@/util/data';
 import SimpleController from '@/controllers/simpleController';
 import Subrace from '@/models/2014/subrace';
 import Trait from '@/models/2014/trait';
-
-// --- Zod Schema ---
-const ParamsSchema = z.object({
-  index: z.string().min(1),
-});
-// --- End Zod Schema ---
 
 const simpleController = new SimpleController(Subrace);
 
@@ -22,7 +16,7 @@ export const show = async (req: Request, res: Response, next: NextFunction) =>
 export const showTraitsForSubrace = async (req: Request, res: Response, next: NextFunction) => {
   try {
     // Validate path parameters
-    const validatedParams = ParamsSchema.safeParse(req.params);
+    const validatedParams = ShowParamsSchema.safeParse(req.params);
     if (!validatedParams.success) {
       return res
         .status(400)
@@ -51,7 +45,7 @@ export const showProficienciesForSubrace = async (
 ) => {
   try {
     // Validate path parameters
-    const validatedParams = ParamsSchema.safeParse(req.params);
+    const validatedParams = ShowParamsSchema.safeParse(req.params);
     if (!validatedParams.success) {
       return res
         .status(400)
